@@ -17,19 +17,19 @@
         }
     ],
     "require": {
-        "php": ">=7.1.3",
-        "doctrine/doctrine-bundle": "^1.8 || ^2.0",
-        "sonata-project/admin-bundle": "^3.63",
-        "sonata-project/doctrine-orm-admin-bundle": "^3.10",
-        "symfony/config": "^4.3",
-        "symfony/framework-bundle": "^4.3"
+        "php": ">=8.2",
+        "doctrine/doctrine-bundle": "^2.10.2",
+        "sonata-project/admin-bundle": "^4.9",
+        "sonata-project/doctrine-orm-admin-bundle": "^4.0",
+        "symfony/config": "^6.2",
+        "symfony/framework-bundle": "^6.2"
     },
     "autoload": {
         "psr-4": { "Prodigious\\Sonata\\MenuBundle\\": "" }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "3.x-dev"
+            "dev-master": "4.x-dev"
         }
     }
 }
