@@ -2,12 +2,12 @@
 
 namespace Prodigious\Sonata\MenuBundle\Admin;
 
-use Sonata\AdminBundle\Admin\AbstractAdmin;
-use Sonata\AdminBundle\Datagrid\ListMapper;
-use Sonata\AdminBundle\Datagrid\DatagridMapper;
-use Sonata\AdminBundle\Form\FormMapper;
-use Sonata\AdminBundle\Route\RouteCollection;
 use Prodigious\Sonata\MenuBundle\Model\MenuInterface;
+use Sonata\AdminBundle\Admin\AbstractAdmin;
+use Sonata\AdminBundle\Datagrid\DatagridMapper;
+use Sonata\AdminBundle\Datagrid\ListMapper;
+use Sonata\AdminBundle\Form\FormMapper;
+use Sonata\AdminBundle\Route\RouteCollectionInterface;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 
 class MenuAdmin extends AbstractAdmin
@@ -17,7 +17,7 @@ class MenuAdmin extends AbstractAdmin
     /**
      * {@inheritdoc}
      */
-    protected function configureFormFields(FormMapper $formMapper)
+    protected function configureFormFields(FormMapper $formMapper): void
     {
         $formMapper
             ->with('config.label_menu', ['translation_domain' => 'ProdigiousSonataMenuBundle'])
@@ -44,7 +44,7 @@ class MenuAdmin extends AbstractAdmin
     /**
      * {@inheritdoc}
      */
-    protected function configureListFields(ListMapper $listMapper)
+    protected function configureListFields(ListMapper $listMapper): void
     {
         $listMapper
             ->add('id', null, ['label' => 'config.label_id', 'translation_domain' => 'ProdigiousSonataMenuBundle'])
@@ -66,7 +66,7 @@ class MenuAdmin extends AbstractAdmin
     /**
      * {@inheritdoc}
      */
-    protected function configureDatagridFilters(DatagridMapper $datagridMapper)
+    protected function configureDatagridFilters(DatagridMapper $datagridMapper): void
     {
         $datagridMapper
             ->add('name')
@@ -77,7 +77,7 @@ class MenuAdmin extends AbstractAdmin
     /**
      * {@inheritdoc}
      */
-    protected function configureRoutes(RouteCollection $collection)
+    protected function configureRoutes(RouteCollectionInterface $collection): void
     {
         $collection->add('items', $this->getRouterIdParameter().'/items');
     }
@@ -85,7 +85,7 @@ class MenuAdmin extends AbstractAdmin
     /**
      * {@inheritdoc}
      */
-    public function configure()
+    protected function configure(): void
     {
         $this->setTemplate('edit', '@ProdigiousSonataMenu/CRUD/edit.html.twig');
     }
@@ -93,7 +93,7 @@ class MenuAdmin extends AbstractAdmin
     /**
      * {@inheritdoc}
      */
-    public function toString($object)
+    public function toString(object $object): string
     {
         return $object instanceof MenuInterface ? $object->getName() : $this->getTranslator()->trans("config.label_menu", array(), 'ProdigiousSonataMenuBundle');
     }
@@ -101,7 +101,7 @@ class MenuAdmin extends AbstractAdmin
     /**
      * @inheritdoc
      */
-    public function prePersist($object)
+    protected function prePersist(object $object): void
     {
         parent::prePersist($object);
         foreach ($object->getMenuItems() as $menuItem) {
@@ -112,7 +112,7 @@ class MenuAdmin extends AbstractAdmin
     /**
      * @inheritdoc
      */
-    public function preUpdate($object)
+    protected function preUpdate(object $object): void
     {
         parent::prePersist($object);
         foreach ($object->getMenuItems() as $menuItem) {
