@@ -3,14 +3,37 @@
 namespace Prodigious\Sonata\MenuBundle\Controller;
 
 use Prodigious\Sonata\MenuBundle\Model\MenuItemInterface;
+use Prodigious\Sonata\MenuBundle\Manager\MenuManager;
+use Sonata\AdminBundle\Admin\Pool;
 use Sonata\AdminBundle\Controller\CRUDController as Controller;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
+use Sonata\AdminBundle\Route\DefaultRouteGenerator;
 use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Contracts\Translation\TranslatorInterface;
+use Doctrine\ORM\EntityManagerInterface;
 
 class MenuItemController extends Controller
 {
 
+    private $menuManager;
+    private $translator;
+    private $routeGenerator;
+    private $adminPool;
+    private $entityManager;
+
+    public function __construct(MenuManager $menuManager,
+                                TranslatorInterface $translator,
+                                DefaultRouteGenerator $routeGenerator,
+                                Pool $adminPool,
+                                EntityManagerInterface $entityManager)
+    {
+        $this->menuManager = $menuManager;
+        $this->translator = $translator;
+        $this->routeGenerator = $routeGenerator;
+        $this->adminPool = $adminPool;
+        $this->entityManager = $entityManager;
+    }
     /**
      * @param integer $id
      */
@@ -26,13 +49,12 @@ class MenuItemController extends Controller
 
         $object->setEnabled(!$object->getEnabled());
 
-        $m = $this->get('doctrine.orm.entity_manager');
+        $m = $this->entityManager;
         $m->persist($object);
         $m->flush();
 
-        return new RedirectResponse($this->get('sonata.admin.route.default_generator')
-            ->generateUrl(
-                $this->get('prodigious_sonata_menu.admin.menu'),
+        return new RedirectResponse($this->routeGenerator->generateUrl(
+            $this->adminPool->getAdminByAdminCode('prodigious_sonata_menu.admin.menu'),
                 'items',
                 ['id' => $object->getMenu()->getId()]
             )
@@ -42,12 +64,11 @@ class MenuItemController extends Controller
     /**
      * {@inheritdoc}
      */
-    protected function redirectTo($object)
+    protected function redirectTo(Request $request, object $object): RedirectResponse
     {
-        $request = $this->getRequest();
-        $response = parent::redirectTo($object, $request);
+        $response = parent::redirectTo($request, $object);
 
-        if (null !== $request->get('btn_update_and_list') || null !== $request->get('btn_create_and_list') || null !== $request->get('btn_update_and_edit') || $this->getRestMethod() === 'DELETE') {
+        if (null !== $request->get('btn_update_and_list') || null !== $request->get('btn_create_and_list') || null !== $request->get('btn_update_and_edit') || strtoupper($request->getMethod()) === 'DELETE') {
             $url = $this->admin->generateUrl('list');
 
             if(!empty($object) && $object instanceof MenuItemInterface) {
@@ -63,4 +84,6 @@ class MenuItemController extends Controller
 
         return $response;
     }
+
+
 }
